requirements.txt
aiohttp==3.8.6
SQLAlchemy==2.0.23
psycopg2-binary==2.9.7
pandas==2.1.3
schedule==1.2.0
backoff==2.2.1
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
loguru==0.7.2
alembic==1.12.1
flake8==6.1.0
black==23.11.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0